added a test function for a shift of -1

+++ rust/profligate/src/caesar.rs
@@ -9,7 +9,7 @@ pub enum CaesarError {
     NonAscii,
 }
 /// Encrypt a string of text with the given key. This operation happens in place to cut down on allocations.
-pub fn encrypt(text: &mut str, key: u8) -> Result<(), CaesarError> {
+pub fn encrypt(text: &mut str, key: i8) -> Result<(), CaesarError> {
     if text.chars().find(|c| !c.is_ascii()).is_some() {
         return Err(CaesarError::NonAscii);
     }
@@ -17,15 +17,15 @@ pub fn encrypt(text: &mut str, key: u8) -> Result<(), CaesarError> {
         // We've already errored out if the text isn't ASCII, so manipulating its bytes in place like a C *char should be safe.
         text.as_bytes_mut()
     };
-    let shift = key % ALPHABET_LENGTH;
+    let shift = key % ALPHABET_LENGTH as i8;
     dbg!(shift);
     for byte in bytes {
         match byte {
             b'a'..=b'z' => {
-                *byte = ((*byte + shift) % b'a') + b'a';
+                *byte = ((*byte as i8 + shift) as u8 % b'a') + b'a';
             }
             b'A'..=b'Z' => {
-                *byte = ((*byte + shift) % b'A') + b'A';
+                *byte = ((*byte as i8 + shift) as u8 % b'A') + b'A';
             }
             _ => (),
         };
